Revert "made 4 working cucumber tests 1 doestnt work"
This reverts commit e7ac6d9141e11cbd17c6590cec034786e2e00bed.

+++ src/test/java/nl/inholland/bankAppBackEnd/stepdefinitions/RegisterStepDefs.java
@@ -3,15 +3,9 @@ package nl.inholland.bankAppBackEnd.stepdefinitions;
 import io.cucumber.java.After;
 import io.cucumber.java.en.*;
 import nl.inholland.bankAppBackEnd.models.User;
-import nl.inholland.bankAppBackEnd.models.BankAccount;
-import nl.inholland.bankAppBackEnd.models.Transaction;
 import nl.inholland.bankAppBackEnd.services.UserService;
-import nl.inholland.bankAppBackEnd.repository.BankAccountRepository;
-import nl.inholland.bankAppBackEnd.repository.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -21,12 +15,6 @@ public class RegisterStepDefs {
     @Autowired
     private UserService userService;
 
-    @Autowired
-    private BankAccountRepository bankAccountRepository;
-
-    @Autowired
-    private TransactionRepository transactionRepository;
-
     private String regUsername = "cucumberuser";
     private String regPassword = "cucumberpass";
     private String regEmail = "cucumber@example.com";
@@ -34,15 +22,7 @@ public class RegisterStepDefs {
 
     @Given("the user is on the registration page")
     public void user_on_registration_page() {
-        // Clean up any existing user and their accounts before registering
         userService.getUserByUsername(regUsername).ifPresent(u -> {
-            List<BankAccount> accounts = bankAccountRepository.findAllByOwner(u);
-            for (BankAccount account : accounts) {
-                List<Transaction> transactions = transactionRepository.findByFromAccountIdOrToAccountIdOrderByTimestampDesc(account.getId(), account.getId());
-                transactionRepository.deleteAll(transactions);
-            }
-            bankAccountRepository.deleteAll(accounts);
-            userService.deleteUserByUsername(regUsername);
         });
     }
 
@@ -74,14 +54,8 @@ public class RegisterStepDefs {
 
     @After
     public void cleanupRegisteredUser() {
-        userService.getUserByUsername(regUsername).ifPresent(u -> {
-            List<BankAccount> accounts = bankAccountRepository.findAllByOwner(u);
-            for (BankAccount account : accounts) {
-                List<Transaction> transactions = transactionRepository.findByFromAccountIdOrToAccountIdOrderByTimestampDesc(account.getId(), account.getId());
-                transactionRepository.deleteAll(transactions);
-            }
-            bankAccountRepository.deleteAll(accounts);
+        if (regUsername != null) {
             userService.deleteUserByUsername(regUsername);
-        });
+        }
     }
 }
